requirements.txt
Pillow>=10.0.0
pyfiglet==0.8.post1
opencv-python-headless>=4.8.0
reflex>=0.3.6
numpy>=1.24.0